fix: Prisma provider auto-switching + PostgreSQL-compatible SQL + Edge Runtime fix
CRITICAL FIX: Production 500 errors caused by Prisma provider mismatch
- Schema had provider=sqlite but Vercel uses PostgreSQL (Neon)
- Prisma client for SQLite rejects postgresql:// URLs

Changes:
- Added scripts/prisma-provider.js: auto-switches provider based on DATABASE_URL
  - file: URLs → sqlite (local dev)
  - postgresql:// URLs → postgresql (Vercel/production)
- Updated build/db scripts to run provider switcher before prisma generate
- Fixed all raw SQL queries for cross-database compatibility:
  - CAST(x AS REAL) → CAST(x AS NUMERIC) (works on both)
  - PostgreSQL ::text → CAST(... AS TEXT) (works on both)
  - FOR UPDATE locks → conditional (PostgreSQL only, SQLite auto-locks)
  - pg_advisory_lock → conditional helpers (no-op on SQLite)
  - Boolean comparisons → CAST(field AS INTEGER) = 1 (works on both)
- Removed setInterval from middleware (Edge Runtime incompatible)
  - Replaced with lazy cleanup on rate-limit checks
- Removed temporary debug-db endpoint
- Added scripts/ to ESLint ignore

+++ src/app/api/admin/affiliates/route.ts
@@ -22,10 +22,10 @@ export async function GET(request: NextRequest) {
       ] = await Promise.all([
         db.$queryRaw<
           Array<{ totalAmount: number; cnt: bigint; status: string }>
-        >`SELECT COALESCE(SUM(CAST("commissionAmount" AS REAL)), 0) as totalAmount, COUNT(*) as cnt, status FROM "AffiliateCommission" GROUP BY status`,
+        >`SELECT COALESCE(SUM(CAST("commissionAmount" AS NUMERIC)), 0) as totalAmount, COUNT(*) as cnt, status FROM "AffiliateCommission" GROUP BY status`,
         db.$queryRaw<
           Array<{ totalAmount: number; totalFees: number; cnt: bigint }>
-        >`SELECT COALESCE(SUM(CAST(amount AS REAL)), 0) as totalAmount, COALESCE(SUM(CAST(fee AS REAL)), 0) as totalFees, COUNT(*) as cnt FROM "AffiliateWithdrawal"`,
+        >`SELECT COALESCE(SUM(CAST(amount AS NUMERIC)), 0) as totalAmount, COALESCE(SUM(CAST(fee AS NUMERIC)), 0) as totalFees, COUNT(*) as cnt FROM "AffiliateWithdrawal"`,
         db.user.findMany({
           where: {
             totalAffiliateEarnings: { not: '0' },
